+++ src/main/java/com/universidadeuropea/comia/controller/homeController.java
@@ -1,4 +1,4 @@
-package com.comia.controller;
+package com.universidadeuropea.comia.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -6,13 +6,11 @@ import org.springframework.web.bind.annotation.RequestMethod;
 
 @Controller
 public class homeController {
-	
-	
+
 	@RequestMapping(value = "/home", method = RequestMethod.GET)
 	public String home() {
-		
+
 		return "home";
 	}
-	
- 
+
 }
