+++ src/main/resources/application.properties
@@ -1,4 +1,5 @@
-spring.datasource.url=jdbc:mysql://localhost:1080/ComIA
-spring.datasource.username=root
-spring.datasource.password=root
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+# BASE DE DATOS
+#spring.datasource.url=jdbc:mysql://localhost:1080/ComIA
+#spring.datasource.username=root
+#spring.datasource.password=root
+#spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
